llama.cpp
* Added experimental ggml-numa backend.
* Refactored ggml-cpu and others to support ggml-numa.

NOTE: We need true tensor model parallelism, which is currently missing in llama.cpp in order to properly parallelize inferencing across multiple GPUs. The theoretical speedup is here, but because NUMA1's calculation depends on the results from NUMA 0, no meaningful speedup can be achieved with the current interdependencies of each split.